Moved CarouselImage and added tests
