+++ src/components/CarouselImage/CarouselImage.tsx
