+++ src/components/CarouselImage/index.ts
